requirements.txt
kafka-python
click
blessings